amend editions list sidebar animation
